+++ src/Pfam/Pfam.jl
@@ -15,21 +15,29 @@ using MIToS.Pfam
 """
 module Pfam
 
-  using MIToS.Utils
-  using MIToS.MSA
-  using MIToS.SIFTS
-  using MIToS.PDB
-  using MIToS.Information
-  using PairwiseListMatrices
-  using DataStructures
-  using ROCAnalysis
+using MIToS.Utils
+using MIToS.MSA
+using MIToS.SIFTS
+using MIToS.PDB
+using MIToS.Information
+using PairwiseListMatrices
+using NamedArrays
+using DataStructures
+using ROCAnalysis
 
-  export  Stockholm, downloadpfam, getseq2pdb,
-          msacolumn2pdbresidue, hasresidues,
-          msacontacts, msaresidues,
-          getcontactmasks, AUC
+export  # Download
+        downloadpfam,
+        # PDB
+        Stockholm,
+        getseq2pdb,
+        msacolumn2pdbresidue,
+        hasresidues,
+        msacontacts,
+        msaresidues,
+        getcontactmasks,
+        AUC
 
-  include("download.jl")
-  include("pdb.jl")
+include("Download.jl")
+include("PDB.jl")
 
 end
